+++ src/canvas.c
@@ -3,7 +3,7 @@
 #include "shmem.h"
 #include "canvas.h"
 
-void ConstructCanvas(Canvas* c, size_t width, size_t height) {
+void ConstructCanvas(Canvas* c, unsigned width, unsigned height) {
     c->buffer = (Tuple3*) shmalloc(width * height * sizeof(Tuple3));
     c->canvas_width = width;
     c->canvas_height = height;
@@ -19,8 +19,8 @@ void DeconstructCanvas(Canvas* c) {
     shfree(c->depth_buffer, c->canvas_width * c->canvas_height * sizeof(float));
 }
 
-void WritePixel(Canvas* c, Tuple3 color, size_t x, size_t y, float depth) {
-    size_t location = (c->canvas_width * y) + x;
+void WritePixel(Canvas* c, Tuple3 color, unsigned x, unsigned y, float depth) {
+    unsigned location = (c->canvas_width * y) + x;
  
     if (depth <= c->depth_buffer[location]) {
         c->buffer[location] = color;
@@ -28,7 +28,7 @@ void WritePixel(Canvas* c, Tuple3 color, size_t x, size_t y, float depth) {
     }
 }
 
-void DirectWritePixel(Canvas* c, Tuple3 color, size_t location, float depth) {
+void DirectWritePixel(Canvas* c, Tuple3 color, unsigned location, float depth) {
     if (depth <= c->depth_buffer[location]) {
         c->buffer[location] = color;
         c->depth_buffer[location] = depth;
@@ -45,8 +45,8 @@ void WriteToPPM(Canvas* c, const char* filename) {
     char width_str [sizeof(int) * 8 + 1];
     char height_str [sizeof(int) * 8 + 1];
 
-    sprintf(width_str, "%ld", c->canvas_width);
-    sprintf(height_str, "%ld", c->canvas_height);
+    sprintf(width_str, "%d", c->canvas_width);
+    sprintf(height_str, "%d", c->canvas_height);
 
     fputs(width_str, fp);
     fputc(' ', fp);
@@ -54,10 +54,10 @@ void WriteToPPM(Canvas* c, const char* filename) {
     fputc('\n', fp);
     fputs("255\n", fp);
 
-    for (size_t i = 0; i < c->canvas_height * c->canvas_width; i++) {
-        float red = fmin(c->buffer[i][0] * 255, 255);
-        float green = fmin(c->buffer[i][1] * 255, 255);
-        float blue = fmin(c->buffer[i][2] * 255, 255);
+    for (unsigned i = 0; i < c->canvas_height * c->canvas_width; i++) {
+        float red = fminf(c->buffer[i][0] * 255, 255);
+        float green = fminf(c->buffer[i][1] * 255, 255);
+        float blue = fminf(c->buffer[i][2] * 255, 255);
 
 
         char red_str[4];
